requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==25.1.0
certifi==2025.1.31
chainmap==1.0.3
charset-normalizer==3.4.1
combomethod==1.0.12
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
iniconfig==2.0.0
nulltype==2.3.1
options==1.4.10
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
selenium==4.28.1
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.2.1
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
